Address all PR #2265 review comments
- Move handle_node_status from BlockchainHandler to NetworkHandler
  (fixes routing: NetworkHandler owns /api/v1/node/ prefix)
- Remove /api/v1/network/ and /api/v1/node/ from BlockchainHandler can_handle
- Disable seed restore (from_words doesn't decode mnemonic — returns error)
- Remove unused connect_default import and node_connected state
- Replace hardcoded "testnet" with runtime environment in directory response
- Validate DID hex decode (return setup_required on failure, not silent empty)
- UI seed grid: support 20 or 24 words with clear messaging
- Add reset_local_identity action to bridge
